Add initial main part of script
Here it is boys and girls, the main part of knowbe4.py. Very basic stuff
if you are familar with APIs and Python, but as a lowly Windows sysadmin
 more familiar with PowerShell and some basic scripts this took me a
while to put together.

Heavily inspired by a few different Python API wrappers that I probably
should cite at some point (one is directly in the code).