Add 1-to-1 mapping of KnowBe4 API
This is the big one. Sort of. I have finished coding all of the
functions that map to what is publicly available from KnowBe4,
though there is a tonnnn of repetition.

TODO: Refactor the main KnowBe4 class to get rid of all the
repetition. Possibly split the various functions out in to its
own class.