Add support for optional parameters
Per the official KnowBe4 API docs, I have added the optional parameters
you can use for looking up users, groups, phishing campaigns, and
training enrollments.

Also added support for deleting empty kwargs from the api_call function.
This is left over from some earlier working code and is probably
unneeded. It's late and I don't feel like messing with it anymore and
it works as it is right now so it's staying in for the moment.